imagefile: allow allocating file with maximum size
This is needed when trying to create an image that is as large as the
volume allows. Implementing the necessary retry loop here is better
than doing it in the caller because here we have more knowledge about
what is a "not enough space" error.

A fixed overhead was tried before, but that didn't work in practice
because the overhead varied depending on the size of the volume. For
example, with a 4 GiB ext4 volume the resulting file has to be 164 MiB
smaller.