operator/deployment: Reconcile deployment changes
Changes to Deployment object are possible in any phase of the
deployment. The controller has to reconcile these changes and make sure
the valid changes are effected to it's sub-resources and invalid
changes are rejected by reverting the change from deployment object
stored at api server.

Example invalid changes are:
- change of device manager mode of a running deployment
- change of driver name which conflict with other deployment

Unit tests added to test if the deployment changes are handled properly
by the controller.