Separate "wait VM responds" loop
F-31 VM enables sshd which gets detected in busyloop,
and we continue to scp and ssh on master.
But F-31 disables sshd soon, to make ssh-keygen.
If we hit that window, ssh will fail.
We try to overcome that by separating first wait as loop
over all VMs, should give time to master (which is the only node at
risk) to handle the keygen job before we try ssh again.