scheduler extender + webhook
The two new components should be compatible with all supported
Kubernetes versions, just the installation may be a bit difficult. For
kubeadm and the QEMU cluster a solution has been found, so the new
components are part of deployment and automatic testing even though
they are not enabled by default in the normal deployments.

As an interim solution they improve the situation with pod scheduling,
but a solution built into Kubernetes would be better.