test: fail when deployment file is missing
The original intention was to silently ignore files that aren't
present because they are not needed. However, currently *all* files
are needed and skipping just has the negative effect that for a new
Kubernetes release, "make start" seems to work but doesn't actually
deploy PMEM-CSI. This happens when using latest Clear Linux and
Kubernetes gets updated there.