Put back "fsdax" as part of VG names
The removal of "fsdax" part in VG names which was made
in connection with "drop sector-mode", introduced
potential API break if VG name was created before change
and used with code after change.
It is safer to put back "fsdax" part of VG name,
it remains hard-coded in all names now, but still
indicates correctly that namespaces are in FSDAX mode.