deploy: simplify certificate handling
There's no reason for using the cluster root certificate and for each
node using its own key (the registry is not checking that). We also
don't need to rename the content of the secrets inside the pod specs.

This makes the setup code easier and avoids the (false!) impression
that the deployment needs different keys for each node. The
description of our TLS handling also becomes simpler.