scheduler: ignore volumes with immediate binding
This is an inexpensive check that avoids the unnecessary pod filtering
for those pods where all volumes have been bound already. The downside
is a conceptual "check before use", i.e. the conditions might still
change between pod creation and scheduling. But changing the storage
class is unlikely, so this should be fine.