test: less verbose setup-ca.sh
Most of the output is not useful when invoking the script as part of
the QEMU cluster creation. When a user calls it, they should know
where it puts the output or (when done through the upcoming
setup-ca-kubernetes.sh) it doesn't matter because it is a temporary
directory.

Bumping up to -loglevel=2 gets rid of:

2020/02/28 10:58:00 [INFO] generating a new CA key and certificate from CSR
2020/02/28 10:58:00 [INFO] generate received request
2020/02/28 10:58:00 [INFO] received CSR
2020/02/28 10:58:00 [INFO] generating key: rsa-2048
2020/02/28 10:58:00 [INFO] encoded CSR
2020/02/28 10:58:00 [INFO] signed certificate with serial number
541485849091675498542977223329122404838977620500

Bumping up to -loglevel=3 also suppresses (harmless?!) warnings which
otherwise just confuse users:

2020/02/28 11:07:22 [WARNING] This certificate lacks a "hosts" field. This makes it unsuitable for
websites. For more information see the Baseline Requirements for the Issuance and Management
of Publicly-Trusted Certificates, v.1.1.6, from the CA/Browser Forum (https://cabforum.org);
specifically, section 10.2.3 ("Information Requirements").