test: skip direct mode unit testing on LVM installations
The direct mode device manager currently runs on LVM deployments only
by chance because /sys happens to be mounted read/write, apparently
related to using the host network. Disabling that causes it to be
read-only.

We let NewPmemDeviceManagerNdctl check for that and then skip
testing if that check has failed.