add Prometheus metrics support
The endpoint is exposed on a separate port for now. It would also be
possible to expose the "/metrics" endpoint on the same port as the
mutating webhook and scheduler extender because all of them accept
unknown peers. The internal gRPC communication has to use a different
port because there we check for known peers.

It's kept separate for now only because it's a bit easier to implement
this way.

The version number is exposed as proof of concept. More metrics data
will follow as needed.

Prometheus language bindings are called directly instead of using the
github.com/kubernetes/component-base/metrics wrapper layer because
that layer doesn't seem to offer any significant advantage and just
increases the set of dependencies.