Jenkinsfile: use env-var DOCKER_REGISTRY instead of hard-coding
Docker registry is used via hard-coded uuid value that is specific to
CI jenkins system, thus limiting Jenkinsfile to be functional
in this Jenkins system only where this UUID is defined.
We can make Jenkinsfile more generic by using a variable.
This variable needs to be provided by Jenkins instance.