remove vendor directory
Downloading sources should be fast and reliable enough that we no
longer need to maintain a copy of the sources in the vendor directory.

Some of our tooling still needs it, though. Repeated builds inside of
a Docker container are faster when they use the "vendor" directory
because each container build starts with an empty $GOPATH directory
and thus would have to download sources anew.

To solve this, the "vendor" directory is populated on demand, it's
just not checked into the git repo.