test: move driver deployment into e2e.test
The immediate advantage is that we can get rid of the logic in the
Jenkinsfile which tries to determine whether the running cluster can
be reused. It also becomes shorter.

Retries get removed because we should get to a point where tests are
so reliable that we don't need them.

The long-term advantage is that we can add tests for various kinds of
deploying PMEM-CSI more easily.