pmem-csi-driver: fix idempotency in node CreateVolume
d1c7d0605b3d98269761c1561d11490a42855764 added a check which detects
the existing volume due to the hash collision, but then always bails
out without checking whether that volume is from a previous call with
the same parameters (idempotency).

The corresponding code in controllerserver-master.go is after
cs.getVolumeByName(volumeName), which looks like the right approach.

While at it, the indention is simplified by adding a "return" after
the code which decides to reuse the existing volume.