test: verify that generated files are up-to-date
We don't want to slow down a PR test by downloading the SDK
unnecessarily, so the same "skip if nothing changed" logic is used as
for go.mod.

This is kept generic as much as possible. However, the new
verify-generated.sh needs to know which files are relevant for code
generation, so it's not completely generic.