test: support @ as separator for regular expressions
Separation by space is problematic in the Jenkinsfile because it
requires quoting. @ is not a special character for the shell and
thus is a bit easier to use as separator. It was picked because
it has no special meaning in a regular expression and is not a
character that should appear in test names and thus also shouldn't
be needed as a plain text match.