test: fix KUBECONFIG message after cluster startup
The message didn't include the full path due to a copy-and-paste error
and inconsistencies in our scripts regarding WORK_DIRECTORY
vs. WORKING_DIRECTORY.

Both names aren't good because they are unspecific (what is "the
working directory"?) and potentially confusing (they don't refer to
the _work directory in PMEM-CSI).

Instead, let's use CLUSTER_DIRECTORY because that's what the directory
is.