pmem-state: do not allow tampering of state file data
Driver's state should not be altered outside of the driver. So added an
integrity check to state data. StateManager computes the hash for the
data being saved and stores the generated hash along with the data.

This change is not backward compatible and breaks loading of older state
files which have no hash information.