test: always test with a fresh cluster
After removing the "reuse cluster" logic in
https://github.com/intel/pmem-csi/pull/548, the same cluster was used
throughout all steps, so instead of testing on the desired OS, the
tests just ran again on the initial one.

A potentially running cluster is intentionally torn down before
testing because that way we can avoid the tear down after the last
step. We don't need to do it then because the whole VM will go down.