test: avoid redundant pod logs of terminated containers
Redeploying PMEM-CSI multiple times during an e2e.test run triggered a
problem in Kubernete's podlogs package: it dumped the output of a
terminated container multiple times until the container either got
replaced or removed.

The fix for this is currently pending, so we have to continue using a
fork of Kubernetes. The only change is the fix for podlogs.