operator: validate deployment certificates
Added a check for validating the provided deployment keys and certificates are
valid to run the driver. This is done by starting a server with given registry
certificates and a client to connect to that server with node-controller
certificates. If the connection success then we treat the certificates are
valid.