operator: reload state of a running deployment if not found in cache
In case of operator restart we refresh all the objects of a reconciling
deployment, to handle the case of operator upgrades that results in
changes to running driver resources.

But at the same time we should be able to handle the conflicting changes
done to deployment, during the time of operator restart. To find those
changes we will not have a cached object to represent the exisitng
deployment. So we prepare the deployment by looking into it's running
sub-resources, we use that for finding the diff.