operator: reload state of a running deployment if not found in cache
When reconciling a running/failed deployment after operator restart we might
not have any previous revision of that deployment in our cache so that
we could have a diff. One thing we could do in this case is force
refresh of all sub-objects of that deployment with new revision.

But we might miss any incompatible changes to that deployment on absence
of the operator. So, for finding those changes we could retrieve those
values from pre-deployed driver objects for that deployment.

We choose "refresh" strategy to support the operator upgrades, which
might have changes to sub-objects that are not visible in deployment
API. One example for this is command-line arguments of sidecar
containers.