test: also delete PMEM-CSI deployment object
In one test run, deleting the existing deployment failed. Perhaps the
operator was re-creating objects. To avoid this and to understand it
better, the deployment object gets deleted first and logging gets
extended to clarify what it is that gets deleted, because the name
itself is often ambiguous (all objects in the operator deployment are
called "pmem-csi-operator").