operator: create all sub-resources in the same phase
The behavior was that first secrets were created, then the phase
changed without waiting for anything and then the next set of objects
were created. This is unnecessarily complex and confusing.

Now all objects get created together in the "new" phase and then the
deployment goes straight to "running". When introducing the "is driver
usable" check, it will first have to go into "initializing" and then
to "running" after the check passes.