operator/test: delete operator on end of test
Operator tests are expected to run with new operator per tests, so that
we can ensure a test is not effected by any leftover previous test data.

But test deployment reuses previously running operator. This change
ensues that the operator gets deleted on end of every test.

Also made changes to reuse deploy.Cluster object and adjusted the time
intervals for tests.