test: repair modified or broken deployments
Originally, test/e2e/deploy just prepared the requested deployment and
relied on tests to keep the deployments in that state. But it is
trivial and useful to also allow tests to modify the deployment
objects without restoring the original state, so now the deployment
setup is repeated before each test.

If nothing has changed, the "kubectl apply" invocations turn into a
NOP. The operator CR always gets updated, but then the operator or
some controller underneath it will notice that nothing has changed.

If something has changed, the intended state is restored.

What still isn't supported is to add objects that aren't part of the
deployment. Tests are still responsible for removing those.