operator/test: ensure all sub-objects gets deleted on delete CR
Current delete deployment CR implemented such that it waits till the CR object
gets deleted, but not that the sub-objects of that CR. This is not
reliable and not sure that those sub-objects gets deleted before start
running the next test.

Random test failures are observed in CI as both "shall not allow
multiple deployments with same driver name" and "shall allow multiple
deployments" uses the same deployment name, and the following test fails
with "already exists" error when it tries to create sub-objects, as
those objects created by earlier test might not get deleted yet.

-----
[2020-04-20T07:27:14.305Z] pmem-csi-operator-688c9795f4-cr7dk/pmem-csi-operator@pmem..ker2: I0420 07:27:14.161380       1 deployment_controller.go:179] Creating: "test-deployment-1" of type "rbac.authorization.k8s.io/v1, Kind=ClusterRoleBinding"
[2020-04-20T07:27:14.306Z] pmem-csi-operator-688c9795f4-cr7dk/pmem-csi-operator@pmem..ker2: I0420 07:27:14.165516       1 deployment_controller.go:140] Requeue: true, error: clusterrolebindings.rbac.authorization.k8s.io "test-deployment-1" already exists
------

Setting `DeleteOptions.PropagationPolicy` to "foreground" ensures that
all dependent objects get deleted before deleting the CR.