test: avoid running controller and node drivers on same node
We haven't found a good solution for the networking issues on Clear
Linux (https://github.com/intel/pmem-csi/issues/555) that plague our
periodic builds.

As a workaround, ensuring that the controller runs on the master node
and thus never on the same node as a node driver avoids the
problematic case.