operator: ignore deployments that are getting deleted
If the deployment has already been marked for deletion, the operator
doesn't need to do anything for it.

This solves an issue encountered when RemoveObjects first deleted the
CR and then lost the race with the app controllers such that the
operator pod got restarted: the operator entered the reconcile loop
and re-created the sub-objects that RemoveObjects had already removed.