test: increase timeout and logging in RemoveObjects
At least in some cases, the race with the app controllers was lost
such that the app controller restarted the operator pod. This then
slows down the removal of all objects, so we need a higher timeout.

The additional logging helps to understand why testing doesn't
proceed. It's rate limited to avoid spamming the logs.