operator: make UpdateOrCreate() generic
While updating an existing object we copy/pick metadata from read object
to our new copy so that we could make UpdateOrCreate() type agnostic.
This will not reset the status fields of the existing object as
client.Update() will not reset the status.

This also changes the semantics of Create() call such that it will not
check the existence of the object being created. So that now it returns
an error if the object is already exits. This revealed and issue in unit
tests of reusing existing CSIDriver that was created for a deleted
deployment.

Took the opportunity to make the Create()/Update() more verbose.