deploy: simplify reference yaml deployments
The purpose is to make the files more consistent within itself (-v
vs. --v, ordering of parameters) and with the objects created by the
operator. That will make it easier to compare object content.

The introduction of the PMEM_CSI_DRIVER_NAME env var simplifies
customization because overwriting an env variable is simpler than
search/replace inside the parameters.

Making the controller unprivileged was something that was done in the
operator, but not in the YAML files.