deploy: compile reference yamls into code and customize them
The intention is to use this for testing the objects created by the
operator (short term) and to use this code to generate the objects
that the operator creates (long term).

The files get compiled into code to support that long-term goal.

Because the operator currently only deploys in production mode, the
debug variants are not currently supported. Adding them will be
simple, because they can be customized the same way, we just need to
start from different input files.