operator: validate objects by comparing against reference files
Originally, changes had to be made in three places when updating
deployments:
- in the YAML files
- in the operator
- in the validateDriverDeployment function

Also, only fields for which comparison code had been written in
validateDriverDeployment were checked. That code was incomplete, which
made it possible to let the operator create objects that were not
quite as desired (see missing TerminationMessagePath).

Now the operator objects get checked by comparing *all* fields (with
some explicit exceptions) against objects derived from the YAML files.

For that to work, it was necessary to make the objects created by the
operator more like those in the YAML files. For examples, now the
operator only creates two secrets instead of three, which is also
simpler (less code, less objects).