operator: better coverage of update scenarios
Updating individual fields leads to corner cases that aren't covered
when only testing of all fields together. For example, the change
detection might miss a field and then the operator would erroneously
decide to not do anything. This is hidden when it updates because of
some other field change.

Updating from defaults may also differ from updating from a deployment
with explicit values.

The same test cases are now used both for unit testing and E2E
testing. The polling interval of 10s made the latter run rather
slowly. To uncover that, logging was used instead of ginkgo.By because
the former adds time stamps.