test: run dax volume test also for driver deployed by operator
The dax testsuite is used because it is a simple volume provisioning
test with additional checking that is relevant for PMEM-CSI. Only the
default filesystem gets tested.

While touching the code, registering the tests gets simplified a bit:
there's no reason to have an extra context because we only test with
one driver.