test: enable/disable webhook when switching deployments
The MutatingWebhookConfigurations.admissionregistration.k8s.io is
another kind of object that some driver deployments have (currently,
only the ones not created by the operator).

We must uninstall it when removing those deployments because it might
not work depending on the names used by the operator deployment.

The scheduler extender can remain configured because it is never
invoked when the mutating webhook is gone and thus the special
label that the extender filters for are not set.