pmem-csi-driver: idempotent NodeUnpublishVolume
When NodeUnpublishVolume is called a second time for an ephemeral
volume, we don't have any information about that volume anymore. This
is not an error. Returning one causes Kubernetes to repeat the call
forever and the pod doesn't terminate (seen with Kubernetes 1.17).

Even for persistent volumes this may be a problem, something that is
currently being discussed upstream.

The code also had another problem: if it got interrupted between
unmounting and deleting the volume, the volume never got deleted
because the code returned too early.

Fixes: https://github.com/intel/pmem-csi/issues/625