docs: cross-reference documentation for different releases
We have not release 0.7.x yet, but can already start referring to it.
The explanation of CSI gets shortened to keep the page small. The TOC
is only needed for viewing on GitHub and can come at the end where it
doesn't distract from the other text on the page.