test: disable Kata Containers by default
Even with the "-vmx" override in the Jenkinsfile removed, nested
virtualization with three levels (Azure HyperV -> QEMU (govm) ->
QEMU (Kata Containers)) was not working well enough for Kata
Containers: because the inner VM runs very slowly, there are timeouts
in the communication between kubelet and Kata Containers ("container
create failed: Failed to check if grpc server is working: rpc error:
code = Unavailable desc = transport is closing").

That means that testing with Kata Containers has to be limited to bare
metal. To achive that, it's turned off by default (and thus in the CI,
which only runs on Azure) and has to be enabled with
TEST_KATA_CONTAINERS_VERSION=1.11.0-rc0 or by invoking
test/setup-kata-containers.sh manually.