docs: fix up floating URLs after writing HTML
Wherever we use github.com/intel/pmem-csi/<something> in a kustomize
example, the generated HTML will have
github.com/intel/pmem-csi/<something>?ref=<commit> instead. This
prevents generating HTML that references files which might no longer
exist after making changes in the repo.

However, files that only exist on the target branch won't be found on
the source. This is the lesser evil because the docsite gets builts in
Jenkins to check the PR.