operator: test deployment in parallel
This replaces Gomega+Ginkgo with plain recursive Go tests and testify
for the asserts. The main advantage is that now tests automatically
run in parallel through goroutines, which reduces the runtime from
over 10 minutes (not sure how much because it timed out) to something
over two minutes on my desktop machine.

Logging output continues to be problematic because it is not related
to the test that triggers it.