test: clean up services manually after operator API tests
The tests used to rely on Kubernetes to garbage-collect the objects.
Starting with Kubernetes 1.18, this got stuck or took a long time (not
sure yet) so often that tests stopped passing.

As a workaround until we can rely on Kubernetes again, the deployment
now gets deleted and then all services for it.