test: better logging of waiting for driver
In one CI run, the driver apparently did not come up, but the
one-per-second "wait for PMEM-CSI" messages also just stopped, which
is odd.

With the new approach for logging, status messages are only printed
once per minute (less logging while waiting) and start and end of
waiting are clearly marked with separate, time-stamped messages.