test: restart tests and data loss
This is part of an investigation around "unexpected error while
removing target path" in NodeUnpublishVolume: the target path got lost
during the sudden power loss that the "can mount again" test simulates
which in combination with insufficient idempotency caused
NodeUnpublishVolume to fail.

The comment for v.unpublish was misleading: the volume has not been
unpublished, only unmounted as part of the reboot.

It is less clear whether the directories created for the driver also
got lost. To be on the safe side, sync is called for the entire
filesystem.