test: avoid running controller and node drivers on same node, II
The previous approach (patching the YAML files) doesn't work for the
operator because it has no API for controlling the node affinity. We
may have to add something for cases like this.

Instead, a "pmem-csi.intel.com/controller" label gets introduced which
through a non-configurable anti-affinity repels the controller
pod. This is currently intentionally *not* documented and should be
considered an implementation detail until we decide to make it part of
the API by describing it.

Both deployment methods now use this approach, i.e. this reverts
commit 53c01a241f73cb1575280943c51b9f5b1e8b9768.