operator: deployment names with dots
The deployment name may contain a dot. But then appending a suffix
like -controller leads to invalid object names because the top-level
domain in a name with dots must not have an hyphen.

We could switch to names like "controller.<name>", but that then has
the same problem when name contains a hyphen. It's also unusual.

So instead, dots get replaced with a hyphen in most objects except the
CSIDriver object, which must have exactly the same name as the driver.