test: test operator in kube-system namespace
We need to verify that the PMEM-CSI driver works when deployed through
the operator in a non-default namespace, like "kube-system". We could
use a new deployment name for it, like
operator-system-direct-production, but then overall test runtime
increases even more.

As a compromise between coverage and total runtime,
operator-direct-production gets changed to deploy the operator and
thus the driver in "kube-system".