CI: replace Clear Linux testing with Fedora
The reason is that we have seen several random test failures in our CI
that only occurred on Clear Linux. This is slowing down PMEM-CSI
development and often we don't have the knowledge and/or time to solve
these issues.

Clear Linux was used to cover Kubernetes 1.17. This now gets tested on
Fedora. PR testing now also happens 1.15, which is more useful than
testing on Clear Linux because we cover the full range of supported
Kubernetes releases; this would have caught the recent issue with the
operator on Kubernetes 1.15 much earlier.

What we loose in the cI is testing with crio. This shouldn't matter
that much because (at least conceptually) PMEM-CSI has no dependency
on the container runtime.