enable Prometheus, gather CSI call counts + PMEM stats
The Prometheus integration uses the approach from
https://github.com/helm/charts/issues/22899:
- HTTP for metrics endpoints
- container ports tell Prometheus which containers to scrape
  and how

CSI call counts are the same as in the sidecars. This enables
correlating statistics and ensures that also node-local operations are
captured; kubelet doesn't seem to be instrumented.

Internal communication is instrumented the same way.

PMEM usage statistics are recorded by querying the active device
manager each time the metrics data gets scraped.

The metrics support is enabled unconditionally in the operator and all
pre-generated deployment files and use plain HTTP for the sake of
simplicity. This is based on the rationale that the data itself is
not sensitive and should always be readily available if desired.