build: update dependencies
Moving to Kubernetes 1.19.0 (at this time, as a pre-release) allows us
to get rid of the fork. The go-autorest replacement also no longer
seems necessary.

Another advantage is that the update allows updating of grpc, which is
good for security (it was rather old before) and for metrics
supports (needs the new grpc.ChainUnaryInterceptor).

For go-logr, the currently unreleased master branch has to be used
because controller-runtime uses the deprecated InfoLogger, which was
removed in go-logr 0.2.1 and re-introduced as a type alias in master
because of that breakage.