deploy: simplify kustomize
Each kustomization is supposed to only depend on input files or other
kustomizations. This wasn't true for the Kubernetes 1.16 files, which
depended on generated files. This happened to work when using "make
kustomize" (but only when not using parallel make!) but was confusing
when invoking kustomize manually (as in "input pmem-csi.yaml file
modified, but output doesn't change"...).

Setting common labels was redundant, the base Kubernetes 1.15
kustomizations already have that.