deploy: make csi.sock work on OpenShift 4.5 + SELinux
For some (unknown) reason SELinux got in the way when using
/var/lib/pmem-csi.intel.com/csi.sock. The symptoms were that
node-driver-registrar was unable to connect and thus the CSI driver
never got registered. Disabling SELinux "fixed" this.

When using the more traditional
/var/lib/kubelet/plugins/pmem-csi.intel.com/csi.sock, it worked with
SELinux enabled, so that is what we are using now.

It makes changing the driver name a bit more complicated, though, in
particular when trying to do it via kustomize.