deploy: handle builtin YAMLs with fixed image version
On the release branch, we inject the version into
deploy/kustomize/driver/pmem-csi.yaml, which then ends up in the
bindata after running "make kustomize" on the release branch. Normally
that is never done, but it might become necessary when back-porting a
fix.

Always building the bindata with "canary" as version is tricky. A
simpler solution is to make the code more flexible which replaces the
image version. \w+ did not match v0.7.0, only the "v0" part. \S+
works.