LVM mode: avoid unnecessary restart of container
Unconditionally returning an error after a successful namespace
creation caused the driver to start. The restarted driver then
finds the existing namespace and continues.

Symptom:

    State:          Running
      Started:      Mon, 24 Aug 2020 09:44:08 +0200
    Last State:     Terminated
      Reason:       Completed
      Message:      failed to run driver: failed to create PMEM namespace with size '33285996544' in region 'region0': <nil>
      Exit Code:    0
      Started:      Mon, 24 Aug 2020 09:44:03 +0200
      Finished:     Mon, 24 Aug 2020 09:44:03 +0200
    Ready:          True
    Restart Count:  1

To detect this and future issues like it, E2E testing now checks the
driver for unexpected restarts. In tests that intentionally restart
pods, the pods must be deleted after the test to reset the restart
counter.