LVM: ignore unsuitable regions during startup
Creating a namespace will only work for regions of the right type.
If we were to attempt creating the namespace, the function would fail
and the entire driver startup would be aborted, just to fail again
when tried again.

There's no test for this because we have no way to simulate different
PMEM configurations.