operator-catalog: Integrated into E2E tests
To ensure that the operator deployed via generated catalog works as
intended to deploy the driver.

Extended the E2E deployment framework by introducing a new
'olm-operator' deploy method. For this deployment method the operator is
deployed via OLM.  Operator deployment API tests make use of this and
runs the tests on operator deployed via OLM.