deploy: retry after node port reuse error
Sometimes when quickly deleting a deployment and recreating it,
Kubernetes rejects a service with
   nodePort: Invalid value: 32000: provided port is already allocated

Kubernetes 1.19 might be a bit better in this regard, but it's unclear
whether this really addresses the issue in our case and we also need
to work with older releases, so we emulate the behavior of an operator
and retry until it succeeds.

Because the error might really be permanent, we time out after one
minute.