remove support for Kubernetes 1.15
Kubernetes 1.15 caused extra complexity because CSIInlineVolume was
still alpha and the driver info file had to be different. With 1.19
being released, 1.15 is out of support upstream and we also can drop
support for it.

To avoid excessive file renaming in the future, the oldest Kubernetes
version for which we maintain kustomize bases is now called "base".