build: use debian:buster as base images, non-reproducible builds
The frequent updates of Clear Linux have been a problem, typically
leading to new PMEM-CSI releases for no good reason other than "there
might have been a bug fix". Debian is expected to update less
frequently. We also no longer update automatically.

The approach taken for that is to use the latest release as base,
whatever that is. In other words, we no longer offer reproducible
releases because there is no way to lock down exactly which files are
taken from the upstream distro. Each release is only built once, so
the binary content is still locked down at build time - we just don't
record in the source code what exactly that binary content is. In
practice that should be good enough.

Compared to Clear Linux, the approach for disabling usage of udev has
changed a bit. use_lvmetad is no longer a valid configuration option.