operator: patch deployment status instead of update
Update fails if there is any mismatch in resource revision due to
intermediate updates to that deployment. Then the operator has to get
fresh copy and update status on that revision. Instead, we can use
patch.

The same observed in CI test failure:
```
deployment_controller.go:172] failed to update status "Running" for deployment "pmem-csi-with-defaults": Operation cannot be fulfilled on deployments.pmem-csi.intel.com "pmem-csi-with-defaults": the object has been modified; please apply your changes to the latest version and try again
```