deploy: exclude socat pods from webhook
It has been observed that after killing all PMEM-CSI pods in the
"sanity PMEM-CSI can mount again after reboot" test (added in commit
35c1182baa8), sometimes starting the socat pods takes a long time (=
minutes) because the daemonset controller tried to re-create them, but
failed because webhook was not up yet. Apparently it then went into an
exponential backoff for minutes.