operator: work around fake client status patch issue
Fake client used in tests does not have any specific implementation for
status patching, instead it uses object patching. That write back the
both status and object changes if any. But we should not apply those
changes to our cached object. They are supposed to captured by next
reconcile loop.

To solve this we are passing the object copy.