operator: generate SAN certificates
Using of certificate CommonName is deprecated in Go-1.15.

Made changes to the self-signed certificate generation and client
certificate verification code so that now they work on DNSNames instead
of CommonName.

Whereas the certificate generation using `cfssl` tools in`setup-ca.sh`
script already setting the hostname(s) which is good enough for us.
`cfssl` treats as an error if no Subject set , so keep setting the CN.

Also updated to latest cfssl release - v1.4.1