test: also distinguish between label and name for operator
Commit eb8448c54c072bab4c9792918b846f8b324cb80f introduced a
distinction between a deployment's label and its name (necessary for
version skew testing). The start/stop scripts for the operator should
use the label, not the name.

This is just done for the sake of consistency and clarity. In practice
there shouldn't have been a difference for the operator
tests (yet...).