test: more robust test<->deployment binding
The anonymous function is bound to the loop variables which change
their content. This works with the current Ginkgo v1.12.1
where Describe calls the function immediately, but in Ginkgo v1.14.1
that changes and the function is stored and called much later.

To prepare for that, each function instance must be bound to copies of
the loop variables. Otherwise for example a test for
"operator" runs with the deployment for "direct-testing".