Dockerfile: stop including fio
The long list of dependencies is causing additional work during
releases (maintaining the bill of material...). It was added as
a convenience for users, but that doesn't justify the extra work and
image size, in particular now that we started to include source code
for those extra packages in the image itself.