build: set VERSION only once
The fallback for unset VERSION is in the top-level Makefile. There's
no need to run that shell command more than once, so better use :=
instead of =.

Overriding the VERSION in operator.make affected the "revision" label
that was set for images after tagging v0.8.0: those then still had
"revision=0.7.0".

The operator.make rules have to use different variable names. Because
those are not specific to the operator, they get set up in the
top-level Makefile.

For the sake of clarity, the output during code generation gets
extended for users who are less familiar with what's going on.