operator: revise deployment status API
Revised the Deployment.Status to accommodate the deployment state
conditions and the driver state. Currently, Deployment has 3 conditions
named CertsVerified, CertsReady, and DriverDeployed. It also records the
summary of controller and node driver state, .i.e, number  of nodes the
driver is running. The state The state get refreshed on each time the
reconcile loop run.