docs: fix shell session quoting
`console` copies just the commands and not their output, but doesn't
handle EOF. `ShellSession` copies everything, including EOF. We need
to be careful about using these correctly, otherwise the copy button
doesn't work as intended.

One limitation is that we cannot mix commands with output and commands
with EOF in the same code block.

Fixes: #773