operator: Deploy the operator in non-default namespace
Changes to operator deployment yamls and scripts to deploy all the operator
objects in it's own namespace - 'pmem-csi'. Even though operator is
running in non-default namespace it requires access to generate events
in default namespace.

Also made changes to e2e tests to deploy the operator in non-default
namespace so that it is possible to validate the event generation.