operator: fix and test code generation
An invalid deploy/crd/pmem-csi.intel.com_deployments.yaml file was
checked in as part of commit b4d17343a8b97785f0aab17b4cd25a9561fb4312
and we didn't noticed until after merging the PR because we didn't
have a test for it and the broken file only caused problems later when
trying to update the master branch (working directory not clean).

Adding the test also showed that "operator-generate-bundle" was
broken (used version in the wrong format). However, even with that
fixed it still fails. It's currently not tested because we don't seem
to need it.