test: replace tar with dd in exec util
"tar" shouldn't be a required component in PMEM-CSI production images
just because it is needed for testing. "dd" is a bit more basic and
for example included in the Red Hat ubi8/ubi-minimal image, in
contrast to "tar".

We only ever needed to transfer a single file, so full directory
transfer is something that we can give up in favor of a simpler
replacement.

An additional bonus is that we no longer need a concurrent process for
producing the input, something that has been problematic in the past.