csi-lib-utils v0.9.0
The new release adds support for process_start_time, which we don't
need and don't want because we already get it through the Prometheus
Golang collector.

Worse, if we keep it enabled in the metrics manager, then metrics
collection fails because the two definitions are slightly different,
which Prometheus treats as an error.