deploy: consistent object naming, app.kubernetes.io labels
The objects were named "pmem-csi-<something>" in the YAML files, which
was inconsistent with the objects created by the operator which uses
the driver name with dots replaced by hyphens as prefix. The approach
taken by the operator is better because it allows parallel
installation of the driver, so now the YAML files do the same.  The
drawback is that the names become longer and have a non-obvious
"-intel-com" in them.

The custom "app" label gets replaced with more expressive
app.kubernetes.io
labels (https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/)
and are used now by the scheduler extender to find the right pods.